backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
pypdf==4.3.1

numpy==1.26.4

torch==2.2.2
torchvision==0.17.2

sentence-transformers==2.6.1
transformers==4.39.3

chromadb==0.4.24
posthog==3.5.0

httpx==0.27.0

groq==0.11.0
python-dotenv==1.0.1
pydantic==2.9.2

ragas==0.2.6
langchain-openai==0.2.14
openai==1.59.7